<refactory> : <DetailSearchPage 위젯을 .dart 파일로 변경>
 <수정 사항>
 - search\all wines\all_wines.dart  =>  search\detaillsearch\all wines\all_wines.dart
 - search\my wines\my_wines.dart  =>  search\detaillsearch\my wines\my_wines.dart
 - search\people\people.dart  =>  search\detaillsearch\people\people.dart

<추가 사항>
 -     =>  search\detaillsearch\search_page.dart
 DetailSearchPage를 search_page.dart에서 사용을 하였는데 search_page라는 이름과 다른 windget이름 때문에 소통하는 와중에 햇갈릴 것으로 예상되어
 refactory하게 되었습니다. ****_page.dart에는 ****_page라는 이름에 메인
 widget을 넣는 것이 더 직관적으로 보입니다
