+++ lib/screens/search/detaillsearch/detail_search_page.dart
@@ -1,0 +1,71 @@
+import 'package:flutter/material.dart';
+import 'all wines/all_wines.dart';
+import 'my wines/my_wines.dart';
+import 'people/people.dart';
+
+class DetailSearchPage extends StatelessWidget {
+  DetailSearchPage({super.key});
+  final TextEditingController _searchController = TextEditingController();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Container(
+          // Add padding around the search bar
+          padding: const EdgeInsets.symmetric(horizontal: 8.0),
+          // Use a Material design search bar
+          child: TextField(
+            cursorColor: Colors.black,
+            controller: _searchController,
+            decoration: InputDecoration(
+              filled: true,
+              fillColor: Colors.white,
+              hintText: 'Search...',
+              // Add a clear button to the search bar
+              suffixIcon: IconButton(
+                icon: Icon(
+                  Icons.clear,
+                  color: Colors.black,
+                ),
+                onPressed: () => _searchController.clear(),
+              ),
+              // Add a search icon or button to the search bar
+              prefixIcon: IconButton(
+                icon: Icon(
+                  Icons.search,
+                  color: Colors.black,
+                ),
+                onPressed: () {
+                  // Perform the search here
+                },
+              ),
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(20.0),
+              ),
+            ),
+          ),
+        ),
+      ),
+      body: DefaultTabController(
+          length: 3,
+          child: Column(
+            children: [
+              TabBar(
+                labelColor: Colors.black,
+                unselectedLabelColor: Colors.grey,
+                tabs: [
+                  Tab(text: 'All Wines'),
+                  Tab(text: 'My Wines'),
+                  Tab(text: 'People'),
+                ],
+              ),
+              Expanded(
+                child: const TabBarView(
+                    children: [AllWines(), MyWines(), People()]),
+              )
+            ],
+          )),
+    );
+  }
+}
