+++ lib/screens/search/detaillsearch/people/people.dart
